move accordion locally
